+++ src/game.rs
@@ -318,7 +318,7 @@ impl Game {
         }
     }
 
-    pub fn start(&mut self, f : fn(&board::Board, usize) -> Option<(f32, node::Node)>, depth : usize) -> Result<(), String> {
+    pub fn start(&mut self, f : fn(&board::Board, u8) -> Option<(f32, node::Node)>, depth : u8) -> Result<(), String> {
         loop {
             // show
             // self.ban.put();
@@ -340,7 +340,7 @@ impl Game {
             self.ban = ban;
 
             // save to kifu
-            self.kifu.append(x, y, teban, rfen);
+            self.kifu.append(x as usize, y as usize, teban, rfen);
 
             // check finished
             if self.ban.is_passpass() {
@@ -361,7 +361,9 @@ impl Game {
         Ok(())
     }
 
-    pub fn start_against_stdin(&mut self, f : fn(&board::Board, usize) -> Option<(f32, node::Node)>, depth : usize, turnin : i8) -> Result<(), String> {
+    pub fn start_against_stdin(&mut self,
+            f : fn(&board::Board, u8) -> Option<(f32, node::Node)>,
+            depth : u8, turnin : i8) -> Result<(), String> {
         loop {
             let x;
             let y;
@@ -380,7 +382,7 @@ impl Game {
                         let mut txt = String::new();
                         io::stdin().read_line(&mut txt).unwrap();
                         txt.pop();
-                        let xx = "0abcdefgh".find(txt.chars().nth(0).unwrap()).unwrap_or(10);
+                        let xx = "0abcdefgh".find(txt.chars().nth(0).unwrap()).unwrap_or(10) as u8;
                         let yy = txt.chars().nth(1).unwrap().to_digit(10);
                         if yy.is_none() {
                             println!("invalid position : {}", txt);
@@ -391,10 +393,10 @@ impl Game {
                             println!("invalid position : {}", txt);
                             continue;
                         }
-                        let pos = (xx, yy as usize);
+                        let pos = (xx, yy as u8);
                         if movable.contains(&pos) {
                             x = xx;
-                            y = yy as usize;
+                            y = yy as u8;
                             break;
                         }
                         println!("{} is not allowed.", txt);
@@ -420,7 +422,7 @@ impl Game {
             self.ban = ban;
 
             // save to kifu
-            self.kifu.append(x, y, teban, rfen);
+            self.kifu.append(x as usize, y as usize, teban, rfen);
 
             // check finished
             if self.ban.is_passpass() {
@@ -441,8 +443,9 @@ impl Game {
         Ok(())
     }
 
-    pub fn start_against_edax(&mut self, f : fn(&board::Board, usize) -> Option<(f32, node::Node)>, depth : usize, turnin : i8)
-            -> Result<(), String> {
+    pub fn start_against_edax(&mut self,
+            f : fn(&board::Board, u8) -> Option<(f32, node::Node)>,
+            depth : u8, turnin : i8) -> Result<(), String> {
         loop {
             let x;
             let y;
@@ -490,9 +493,9 @@ impl Game {
 
                     // println!("{}{} from EDAX:{}", xtxt, ytxt, lines[2]);
 
-                    x = "0abcdefgh".find(xtxt).unwrap_or(10);
-                    y = ytxt.to_digit(10).unwrap() as usize;
-            }
+                    x = "0abcdefgh".find(xtxt).unwrap_or(10) as u8;
+                    y = ytxt.to_digit(10).unwrap() as u8;
+                }
            } else {
                 println!("{}", self.ban.to_str());
                 // think
@@ -511,7 +514,7 @@ impl Game {
             self.ban = ban;
 
             // save to kifu
-            self.kifu.append(x, y, teban, rfen);
+            self.kifu.append(x as usize, y as usize, teban, rfen);
 
             // check finished
             if self.ban.is_passpass() {
@@ -533,8 +536,8 @@ impl Game {
     }
 
     pub fn start_with_2et(&mut self,
-            f : fn(&board::Board, usize) -> Option<(f32, node::Node)>,
-            depth : usize, et1 : &weight::Weight, et2 : &weight::Weight)
+            f : fn(&board::Board, u8) -> Option<(f32, node::Node)>,
+            depth : u8, et1 : &weight::Weight, et2 : &weight::Weight)
             -> Result<(), String> {
         loop {
             // show
@@ -567,7 +570,7 @@ impl Game {
             self.ban = ban;
 
             // save to kifu
-            self.kifu.append(x, y, teban, rfen);
+            self.kifu.append(x as usize, y as usize, teban, rfen);
 
             // check finished
             if self.ban.is_passpass() {
